Prints turn types and modifiers as strings
