+++ unit_tests/engine/guidance_assembly.cpp
@@ -46,7 +46,7 @@ BOOST_AUTO_TEST_CASE(trim_short_segments)
                                      0.2,
                                      1.9076601161280742,
                                      0.2,
-                                     TRAVEL_MODE_DRIVING,
+                                     extractor::TRAVEL_MODE_DRIVING,
                                      {{FloatLongitude{-73.981492}, FloatLatitude{40.768258}},
                                       329,
                                       348,
@@ -67,7 +67,7 @@ BOOST_AUTO_TEST_CASE(trim_short_segments)
                                      0,
                                      0,
                                      0,
-                                     TRAVEL_MODE_DRIVING,
+                                     extractor::TRAVEL_MODE_DRIVING,
                                      {{FloatLongitude{-73.981495}, FloatLatitude{40.768275}},
                                       0,
                                       0,
