+++ unit_tests/engine/json_factory.cpp
@@ -1,4 +1,5 @@
 #include "engine/api/json_factory.hpp"
+#include "extractor/guidance/turn_instruction.hpp"
 
 #include <boost/test/test_case_template.hpp>
 #include <boost/test/unit_test.hpp>
